requirements.txt
spotipy
python-dotenv
flask
flask-session
